refactor: 리뷰 반영, 수익률 구하는 로직을 LottoResult에 위임
수익률 담당 책임을 LottoResult가 가지도록 수정
- LottoResult에 전체 가격을 계산하는 함수 추가
- Rank 클래스에서 계산하는 함수 제거
- 관련 테스트 코드 수정
